# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Kudesnik. Windows \u0026 Doors | REST-API scheme",
        "title": "Kudesnik API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:5050",
    "basePath": "/api/v1",
    "paths": {
...
        "/products/visibility": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Изменить видимость списка товаров по списку их ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Изменить видимость списка товаров",
                "parameters": [
                    {
                        "description": "UpdateProductsVisability тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_products_transport_http.UpdateProductsVisabilityRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/products/visibility/p/{id}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Изменить видимость товара",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Изменить видимость товара",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID обновляемой заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "UpdateProductVisability тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_products_transport_http.UpdateProductVisabilityRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "internal_features_products_transport_http.ProductIdDTORequest": {
            "type": "object",
            "required": [
                "product_id"
            ],
            "properties": {
                "product_id": {
                    "type": "string"
                }
            }
        },
        "internal_features_products_transport_http.UpdateProductVisabilityRequestDTO": {
            "type": "object",
            "properties": {
                "is_visible": {
                    "type": "boolean"
                }
            }
        },
        "internal_features_products_transport_http.UpdateProductsVisabilityRequestDTO": {
            "type": "object",
            "properties": {
                "is_visible": {
                    "type": "boolean"
                },
                "productsIDs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_features_products_transport_http.ProductIdDTORequest"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Some queries require authorization: \"Bearer \u003cjwt_token\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}